mmc: Add mount services
The iterm updater relies on the versions being mounted in the
MEDIA_DIR in order to create the D-Bus objects when the BMC
first starts up.

Create two services: one to mount the emmc rootfs partitions
so that the version D-Bus objects get created, and another one
to unmount them after the item updater has run because there's
no need to keep the mounts around, actually they shouldn't be
mounted during a code update. The UBI layout does an unmount
at the start of an update, but it's less risky to just unmount
once the mount is no longer needed in case the update is done
manually by starting the update service file or directly
calling obmc-flash-bmc.

Tested: Verified the D-Bus version objects are created after
BMC reboot.

Change-Id: Ic0c83391a2879b8508bf5aa9f18941590d96352d
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>